NumericTrack class and methods

The virtual parent class for all track items in the Gviz package designed to contain numeric data. This class merely exists for dispatching purpose.

# S4 method for class 'NumericTrack'
drawAxis(GdObject, from, to, ...)

# S4 method for class 'NumericTrack'
drawGrid(GdObject, from, to)

Arguments

GdObject

Object of class GdObject.

from, to

Numeric scalar, giving the range of genomic coordinates to limit the tracks in. Note that from cannot be larger than to.

...

Additional arguments.

Value

A virtual class: No objects may be created from it.

Functions

drawAxis(NumericTrack): add a y-axis to the title panel of a track.

drawGrid(NumericTrack): superpose a grid on top of a track.

Slots

dp

Object of class DisplayPars, the display settings controlling the look and feel of a track. See settings for details on setting graphical parameters for tracks.

name

Object of class character, a human-readable name for the track that will be used in the track's annotation panel if necessary.

imageMap

Object of class ImageMap, containing optional information for an HTML image map. This will be created by the drawGD methods when the track is plotted to a device and is usually not set by the user.

range

Object of class GRanges, the genomic ranges of the track items as well as additional annotation information in its elementMetadata slot. Please note that the slot is actually implemented as a class union between GRanges and IRanges to increase efficiency, for instance for DataTrack objects. This usually does not concern the user.

chromosome

Object of class character, the chromosome on which the track is defined. There can only be a single chromosome for one track. For certain subclasses, the space of allowed chromosome names is limited (e.g., only those chromosomes that exist for a particular genome). Throughout the package, chromosome names have to be entered either as a single integer scalar or as a character scalar of the form chrXYZ, where XYZ may be an arbitrary character string.

genome

Object of class character, the genome for which the track is defined. For most sub-classes this has to be a valid UCSC genome identifier, however this may not always be formally checked upon object instantiation.

See also

Author

Florian Hahne