Export Gviz tracks into an annotation file representation.

This function is still a bit experimental. Write all tracks provided as a list tracks into a single BED file. So far only BED export is supported.

exportTracks(tracks, range, chromosome, file)

Arguments

tracks

A list of annotation track objects to be exported into a single BED file.

range

A numeric vector or length 2. The genomic range to display when opening the file in a browser.

chromosome

The chromosome to display when opening the file in a browser.

file

Character, the path to the file to write into.

Value

The function is called for its side effect of writing to a file.

Author

Florian Hahne

Examples

## export AnnotationTrack to BED file
at <- AnnotationTrack(start = seq(1, 10), width = 1, chromosome = "chr1")
exportTracks(list(at),
    range = c(1, 10), chromosome = "chr1",
    file = paste0(tempfile(), ".bed")
)