HighlightTrack class and methods

A container for other track objects from the Gviz package that allows for the addition of a common highlighting area across tracks.

HighlightTrack(trackList=list(), range=NULL, start=NULL, end=NULL, width=NULL, 
               chromosome, genome, name="HighlightTrack",  ...)

Arguments

trackList
A list of Gviz track objects that all have to inherit from class GdObject.

range
An optional meta argument to handle the different input types. If the range argument is missing, all the relevant information to create the object has to be provided as individual function arguments (see below).

The different input options for range are:

A GRanges object: the genomic ranges for the highlighting regions.

An IRanges object: almost identical to the GRanges case, except that the chromosome information has to be provided in the separate chromosome argument, because it can not be directly encoded in an IRanges object.

A data.frame object: the data.frame needs to contain at least the two mandatory columns start and end with the range coordinates. It may also contain a chromosome column with the chromosome information for each range. If missing, this information will be drawn from the constructor's chromosome argument.

start, end
An integer scalar with the genomic start or end coordinates for the highlighting range. Can also be supplied as part of the range argument.

width
An integer vector of widths for highlighting ranges. This can be used instead of either start or end to specify the range coordinates.

chromosome
The chromosome on which the track's genomic ranges are defined. A valid UCSC chromosome identifier if options(ucscChromosomeNames=TRUE). Please note that in this case only syntactic checking takes place, i.e., the argument value needs to be an integer, numeric character or a character of the form chrx, where x may be any possible string. The user has to make sure that the respective chromosome is indeed defined for the the track's genome. If not provided here, the constructor will try to build the chromosome information based on the available inputs, and as a last resort will fall back to the value chrNA. Please note that by definition all objects in the Gviz package can only have a single active chromosome at a time (although internally the information for more than one chromosome may be present), and the user has to call the chromosome<- replacement method in order to change to a different active chromosome.

genome
The genome on which the track's ranges are defined. Usually this is a valid UCSC genome identifier, however this is not being formally checked at this point. If not provided here the constructor will try to extract this information from the provided inputs, and eventually will fall back to the default value of NA.

name
Character scalar of the track's name. This is not really used and only exists fro completeness.

...
All additional parameters are ignored.

Objects from the Class

Objects can be created using the constructor function HighlightTrack.

Details

A track to conceptionally group other Gviz track objects into a meta track for the sole purpose of overlaying all the contained tracks with the same highlighting region as defined by the objects genomic ranges. During rendering the contained tracks will be treated as if they had been provided to the plotTracks function as individual objects.

Value

The return value of the constructor function is a new object of class HighlightTrack.

Slots

trackList:

Object of class "list", holding the subtrack objects.

range:

Object of class GRanges, inherited from class RangeTrack

chromosome:

Object of class "character", inherited from class RangeTrack

genome:

Object of class "character", inherited from class RangeTrack

dp:

Object of class DisplayPars, inherited from class GdObject

name:

Object of class "character", inherited from class GdObject

imageMap:

Object of class ImageMap, inherited from class GdObject

Extends

Class "RangeTrack", directly.

Class "GdObject", by class "RangeTrack", distance 2.

Methods

In the following code chunks, obj is considered to be an object of class HighlightTrack.

Internal methods:

setStacks

signature(GdObject="HighlightTrack"): recompute the stacks based on the available space and on the object's track items and stacking settings. This really just calls the setStacks methods for the contained tracks and only exists for dispatching reasons. Usage: setStacks(GdObject, ...) Examples:

Gviz:::setStacks(obj)

initialize

signature(.Object="HighlightTrack"): initialize the object.

subset

signature(x="HighlightTrack"): subset all the contained tracks in an HighlightTrack by coordinates and sort if necessary. Usage: subset(x, ...) Additional Arguments:

...: additional arguments are passed on to the next methods.

Examples:

subset(obj)

length

signature(x="HighlightTrack"): return the number of items in the track. Usage: length(x) Examples:

length(obj)

Inherited methods:

[

signature(x="HighlightTrack", i="ANY", j="ANY", drop="ANY"): subset the items in the HighlightTrack object. This is essentially similar to subsetting of the GRanges object in the range slot. For most applications, the subset method may be more appropriate. Additional Arguments:

i, j: subsetting indices, j is ignored.

drop: argument is ignored.

Examples:

obj[1:5]

chromosome

signature(GdObject="HighlightTrack"): return the chromosome for which the track is defined. Usage: chromosome(GdObject) Examples:

chromosome(obj)

chromosome<-

signature(GdObject="HighlightTrack"): replace the value of the track's chromosome. This has to be a valid UCSC chromosome identifier or an integer or character scalar that can be reasonably coerced into one. Usage: chromosome<-(GdObject, value) Additional Arguments:

value: replacement value.

Examples:

chromosome(obj) <- "chr12"

start, end, width

signature(x="HighlightTrack"): the start or end coordinates of the track items, or their width in genomic coordinates. Usage: start(x) end(x) width(x) Examples:

start(obj)

end(obj)

width(obj)

start<-, end<-, width<-

signature(x="HighlightTrack"): replace the start or end coordinates of the track items, or their width. Usage: start<-(x, value) end<-(x, value) width<-(x, value) Additional Arguments:

value: replacement value.

Examples:

start(obj) <- 1:10

end(obj) <- 20:30

width(obj) <- 1

position

signature(GdObject="HighlightTrack"): the arithmetic mean of the track item's coordionates, i.e., (end(obj)-start(obj))/2. Usage: position(GdObject) Examples:

position(obj)

genome

signature(x="HighlightTrack"): return the track's genome. Usage: genome(x) Examples:

genome(obj)

genome<-

signature(x="HighlightTrack"): set the track's genome. Usually this has to be a valid UCSC identifier, however this is not formally enforced here. Usage: genome<-(x, value) Additional Arguments:

value: replacement value.

Examples:

genome(obj) <- "mm9"

range

signature(x="HighlightTrack"): return the genomic coordinates for the track as an object of class IRanges. Usage: range(x) Examples:

range(obj)

ranges

signature(x="HighlightTrack"): return the genomic coordinates for the track along with all additional annotation information as an object of class GRanges. Usage: ranges(x) Examples:

ranges(obj)

split

signature(x="HighlightTrack"): split a HighlightTrack object by an appropriate factor vector (or another vector that can be coerced into one). The output of this operation is a list of objects of the same class as the input object, all inheriting from class HighlightTrack. Usage: split(x, f, ...) Additional Arguments:

f: the splitting factor.

...: all further arguments are ignored.

Examples:

split(obj, c("a", "a", "b", "c", "a"))

displayPars

signature(x="HighlightTrack", name="character"): list the value of the display parameter name. See settings for details on display parameters and customization. Usage: displayPars(x, name) Examples:

displayPars(obj, "col")

displayPars

signature(x="HighlightTrack", name="missing"): list the value of all available display parameters. See settings for details on display parameters and customization. Examples:

displayPars(obj)

getPar

signature(x="HighlightTrack", name="character"): alias for the displayPars method. See settings for details on display parameters and customization. Usage: getPar(x, name) Examples:

getPar(obj, "col")

getPar

signature(x="HighlightTrack", name="missing"): alias for the displayPars method. See settings for details on display parameters and customization. Examples:

getPar(obj)

displayPars<-

signature(x="HighlightTrack", value="list"): set display parameters using the values of the named list in value. See settings for details on display parameters and customization. Usage: displayPars<-(x, value) Examples:

displayPars(obj) <- list(col="red", lwd=2)

setPar

signature(x="HighlightTrack", value="character"): set the single display parameter name to value. Note that display parameters in the HighlightTrack class are pass-by-reference, so no re-assignmnet to the symbol obj is necessary. See settings for details on display parameters and customization. Usage: setPar(x, name, value) Additional Arguments:

name: the name of the display parameter to set.

Examples:

setPar(obj, "col", "red")

setPar

signature(x="HighlightTrack", value="list"): set display parameters by the values of the named list in value. Note that display parameters in the HighlightTrack class are pass-by-reference, so no re-assignmnet to the symbol obj is necessary. See settings for details on display parameters and customization. Examples:

setPar(obj, list(col="red", lwd=2))

names

signature(x="HighlightTrack"): return the value of the name slot. Usage: names(x) Examples:

names(obj)

names<-

signature(x="HighlightTrack", value="character"): set the value of the name slot. Usage: names<-(x, value) Examples:

names(obj) <- "foo"

coords

signature(ImageMap="HighlightTrack"): return the coordinates from the internal image map. Usage: coords(ImageMap) Examples:

coords(obj)

tags

signature(x="HighlightTrack"): return the tags from the internal image map. Usage: tags(x) Examples:

tags(obj)

Display Parameters

The following display parameters are set for objects of class HighlightTrack upon instantiation, unless one or more of them have already been set by one of the optional sub-class initializers, which always get precedence over these global defaults. See settings for details on setting graphical parameters for tracks.

col="red": Integer or character vector. The border color for the highlighting regions.

fill="#FFE3E6": Integer or character vector. The fill color for the highlighting regions.

inBackground=TRUE: Logical scalar. Place the box in the background or in the foreground.

Additional display parameters are being inherited from the respective parent classes. Note that not all of them may have an effect on the plotting of HighlightTrack objects.

GdObject:

alpha=1: Numeric scalar. The transparency for all track items.

alpha.title=NULL: Numeric scalar. The transparency for the title panel.

background.legend="transparent": Integer or character scalar. The background color for the legend.

background.panel="transparent": Integer or character scalar. The background color of the content panel.

background.title="lightgray": Integer or character scalar. The background color for the title panel.

cex=1: Numeric scalar. The overall font expansion factor for all text and glyphs, unless a more specific definition exists.

cex.axis=NULL: Numeric scalar. The expansion factor for the axis annotation. Defaults to NULL, in which case it is automatically determined based on the available space.

cex.title=NULL: Numeric scalar. The expansion factor for the title panel. This effects the fontsize of both the title and the axis, if any. Defaults to NULL, which means that the text size is automatically adjusted to the available space.

col.axis="white": Integer or character scalar. The font and line color for the y axis, if any.

col.border.title="white": Integer or character scalar. The border color for the title panels.

col.frame="lightgray": Integer or character scalar. The line color used for the panel frame, if frame==TRUE

col.grid="#808080": Integer or character scalar. Default line color for grid lines, both when type=="g" in DataTracks and when display parameter grid==TRUE.

col.line=NULL: Integer or character scalar. Default colors for plot lines. Usually the same as the global col parameter.

col.symbol=NULL: Integer or character scalar. Default colors for plot symbols. Usually the same as the global col parameter.

col.title="white" (Aliases: fontcolor.title): Integer or character scalar. The border color for the title panels

collapse=TRUE: Boolean controlling whether to collapse the content of the track to accomodate the minimum current device resolution. See collapsing for details.

fontcolor="black": Integer or character scalar. The font color for all text, unless a more specific definition exists.

fontface=1: Integer or character scalar. The font face for all text, unless a more specific definition exists.

fontface.title=2: Integer or character scalar. The font face for the title panels.

fontfamily="sans": Integer or character scalar. The font family for all text, unless a more specific definition exists.

fontfamily.title="sans": Integer or character scalar. The font family for the title panels.

fontsize=12: Numeric scalar. The font size for all text, unless a more specific definition exists.

frame=FALSE: Boolean. Draw a frame around the track when plotting.

grid=FALSE: Boolean, switching on/off the plotting of a grid.

h=-1: Integer scalar. Parameter controlling the number of horizontal grid lines, see panel.grid for details.

lineheight=1: Numeric scalar. The font line height for all text, unless a more specific definition exists.

lty="solid": Numeric scalar. Default line type setting for all plotting elements, unless there is a more specific control defined elsewhere.

lty.grid="solid": Integer or character scalar. Default line type for grid lines, both when type=="g" in DataTracks and when display parameter grid==TRUE.

lwd=1: Numeric scalar. Default line width setting for all plotting elements, unless there is a more specific control defined elsewhere.

lwd.border.title=1: Integer scalar. The border width for the title panels.

lwd.grid=1: Numeric scalar. Default line width for grid lines, both when type=="g" in DataTracks and when display parameter grid==TRUE.

lwd.title=1: Integer scalar. The border width for the title panels

min.distance=1: Numeric scalar. The minimum pixel distance before collapsing range items, only if collapse==TRUE. See collapsing for details.

min.height=3: Numeric scalar. The minimum range height in pixels to display. All ranges are expanded to this size in order to avoid rendering issues. See collapsing for details.

min.width=1: Numeric scalar. The minimum range width in pixels to display. All ranges are expanded to this size in order to avoid rendering issues. See collapsing for details.

reverseStrand=FALSE: Logical scalar. Set up the plotting coordinates in 3' -> 5' direction if TRUE. This will effectively mirror the plot on the vertical axis.

rotation=0: The rotation angle for all text unless a more specific definiton exists.

rotation.title=90 (Aliases: rotation.title): The rotation angle for the text in the title panel. Even though this can be adjusted, the automatic resizing of the title panel will currently not work, so use at own risk.

showAxis=TRUE: Boolean controlling whether to plot a y axis (only applies to track types where axes are implemented).

showTitle=TRUE: Boolean controlling whether to plot a title panel. Although this can be set individually for each track, in multi-track plots as created by plotTracks there will still be an empty placeholder in case any of the other tracks include a title. The same holds true for axes. Note that the the title panel background color could be set to transparent in order to completely hide it.

size=1: Numeric scalar. The relative size of the track. Can be overridden in the plotTracks function.

v=-1: Integer scalar. Parameter controlling the number of vertical grid lines, see panel.grid for details.

Author

Florian Hahne

See also

Examples

## Object construction:
set.seed(123)
dat <- runif(100, min=-2, max=22)
gt <- GenomeAxisTrack()
dt <- DataTrack(data=dat, start=sort(sample(200, 100)), width=1, genome="hg19")

ht <- HighlightTrack(trackList=list(gt, dt))