ImageMap: HTML image map information

HTML image map information for annotation tracks.

coords(ImageMap, ...)

# S4 method for `NULL`
coords(ImageMap)

# S4 method for ImageMap
coords(ImageMap)

tags(ImageMap, ...)

# S4 method for `NULL`
tags(ImageMap)

# S4 method for ImageMap
tags(ImageMap)

Arguments

ImageMap

Object of ImageMap-class, containing optional information for an HTML image map.

Value

Returns the coordinates from the image map.

Returns the tags from the image map.

Details

Objects of the ImageMap-class are usually not created by the user, hence the constructor function ImageMap is not exported in the name space.

Functions

coords(): Generics for coords.

coords(`NULL`): Returns the coordinates from the image map.

coords(ImageMap): Returns the coordinates from the image map.

tags(): Generics for tags.

tags(`NULL`): Returns the tags from the image map

tags(ImageMap): Returns the tags from the image map

Slots

coords

Object of class matrix, the image map coordinates. In the order x bl, y bl, x tr, y tr. Row names are mandatory for the matrix and have to be unique.

tags

Object of class list, the individual HTML tags for the image map. The value of each list item has to be a named character vector, where the names must match back into the row names of the coords matrix

Examples

## Not provided. This is an internal structure.